Enable -Wextra and -Wformat friends
-Wformat is included in -Wall, which can be removed from the list.
-Wno-unused-parameter is added as main() test/test_runner.c currently
doesn't do anything with the arguments.

+++ server.c
@@ -132,13 +132,14 @@ int server_accept(const struct server_info* server,
 int read_client_event(struct client_info* client, struct client_event* event) {
     uint8_t event_buffer[EV_MSG_SIZE];
 
-    int read_length = read(client->cl_fd, event_buffer, sizeof(event_buffer));
+    ssize_t read_length = read(client->cl_fd, event_buffer,
+            sizeof(event_buffer));
     if (read_length < 0 && errno != EINTR) {
         LOG_ERRNO("error reading from client");
         return -1;
     }
 
-    if (read_length < sizeof(event_buffer)) {
+    if (read_length < (ssize_t)sizeof(event_buffer)) {
         return 0;
     }
